+++ lib/reel.rb
@@ -11,18 +11,16 @@ require 'reel/version'
 require 'reel/mixins'
 require 'reel/connection'
 require 'reel/logger'
+require 'reel/request_info'
 require 'reel/request'
+require 'reel/request_body'
 require 'reel/request_parser'
 require 'reel/response'
+require 'reel/response_writer'
 require 'reel/server'
 require 'reel/ssl_server'
 require 'reel/websocket'
 require 'reel/stream'
-
-require 'rack'
-require 'rack/handler'
-require 'rack/handler/reel'
-require 'reel/rack_worker'
 
 # A Reel good HTTP server
 module Reel