+++ lib/reel/server.rb
@@ -12,6 +12,7 @@ module Reel
       # This is actually an evented Celluloid::IO::TCPServer
 
       @server = TCPServer.new(host, port)
+      @server.setsockopt(Socket::IPPROTO_TCP, Socket::TCP_NODELAY, 1)
       @server.listen(backlog)
       @callback = callback
       async.run