+++ CHANGES.md
@@ -1,8 +1,15 @@
-0.4.0
-----
+0.4.0.pre2
+----------
+* Pipelining support
+* Reel::Connection#each_request for iterating through keep-alive requests
+* Reel::Request#body now returns a Reel::RequestBody object instead of a String
+* New WebSocket API: obtain WebSockets through Reel::Request#websocket instead
+  of through Reel::Connection#request. Allows processing of WebSockets through
+  other means than the built-in WebSocket support
 * Allow Reel to stop cleanly
 * Remove `on_error` callback system
 * Increase buffer size
+* Remove Reel::App (unmaintained, sorry)
 
 0.3.0
 -----